신규 renderScope props 구현 및 구조 개선
