+++ src/provider/DialogProvider.tsx
@@ -1,7 +1,7 @@
 import type { PropsWithChildren } from 'react';
 import { useState } from 'react';
 
-import DialogRenderProvider from '@provider/DialogRenderProvider';
+import DialogRenderProvider from '@provider/DialogRootProvider';
 import DialogStatesContext from '@context/DialogStatesContext';
 
 import type { DialogState } from '@types';
