+++ src/stories/index.stories.tsx
@@ -2,11 +2,10 @@ import { useState } from 'react';
 import type { RefAttributes } from 'react';
 
 import type { Meta, StoryObj } from '@storybook/react';
-import { Button, Typography } from '@mrcamelhub/camel-ui';
+import { Button, Input, Typography } from '@mrcamelhub/camel-ui';
+import Dialog from '@components/Dialog';
 
 import type { DialogProps } from '@types';
-
-import Dialog from '.';
 
 const meta: Meta<typeof Dialog> = {
   title: 'Dialog',
@@ -26,6 +25,16 @@ function DialogWithHooks(args: DialogProps & RefAttributes<HTMLDivElement>) {
       </Button>
       <Dialog {...args} open={open} onClose={() => setOpen(false)}>
         <Typography>Camel Dialog</Typography>
+        {args?.fullScreen && (
+          <Typography
+            weight="bold"
+            customStyle={{
+              marginTop: 8
+            }}
+          >
+            아래로 스와이프해서 닫을 수 있어요!
+          </Typography>
+        )}
       </Dialog>
     </>
   );
@@ -60,13 +69,18 @@ function NestedDialogWithHooks(args: DialogProps & RefAttributes<HTMLDivElement>
           Open Second Dialog
         </Button>
       </Dialog>
-      <Dialog {...args} open={openSecondDialog} onClose={() => setOpenSecondDialog(false)}>
+      <Dialog
+        {...args}
+        open={openSecondDialog}
+        onClose={() => setOpenSecondDialog(false)}
+        renderScope="component"
+      >
         <Typography
           customStyle={{
             marginBottom: 8
           }}
         >
-          Camel Dialog
+          Camel Dialog (Render Scope: component)
         </Typography>
         <Button
           variant="solid"
@@ -86,6 +100,30 @@ function NestedDialogWithHooks(args: DialogProps & RefAttributes<HTMLDivElement>
   );
 }
 
+function RenderScopeDialogWithHooks(args: DialogProps & RefAttributes<HTMLDivElement>) {
+  const [open, setOpen] = useState(false);
+  const [value, setValue] = useState('');
+
+  return (
+    <>
+      <Button variant="solid" brandColor="primary" onClick={() => setOpen(true)}>
+        Open Dialog
+      </Button>
+      <Dialog {...args} open={open} onClose={() => setOpen(false)}>
+        <Typography>Camel Dialog</Typography>
+        <Input
+          onChange={(e) => setValue(e.currentTarget.value)}
+          value={value}
+          placeholder="입력해 주세요."
+          customStyle={{
+            marginTop: 8
+          }}
+        />
+      </Dialog>
+    </>
+  );
+}
+
 export const Default: Story = {
   render: (args) => <DialogWithHooks {...args} />
 };
@@ -96,4 +134,8 @@ export const Nested: Story = {
 
 export const FullScreen: Story = {
   render: (args) => <DialogWithHooks {...args} fullScreen />
+};
+
+export const RenderScope: Story = {
+  render: (args) => <RenderScopeDialogWithHooks {...args} renderScope="component" />
 };
